Launcher3: bring back hotseat icons view for recents ui state
After the addition of prediction row for apps, recents ui got
messed up as the appredictions row gets pushed down
due to the searchbox also showing in the FloatingHeaderView.

On Android 11, this issue is only observed on 2 button navigation

Get rid of the pesky app predictions row and stick with the hotseat
icons view.

Change-Id: I7322aea223e01fbf518f6fdd6aea29b3ff4401d6